VER3-453 now reading custom partial matrices
