+++ app/app/calculation_service/model/isu_factors.py
@@ -6,21 +6,7 @@ from app.calculation_service.model.outcome import Outcome
 from app.calculation_service.model.predictor import Predictor
 from app.calculation_service.model.repeated_measure import RepeatedMeasure
 from app.calculation_service.utilities import list_compare
-
-class ContrastMatrix(object):
-    """Class to store a custom contrast matrix"""
-
-    def __init_(self,
-                matrixType: str = None,
-                values: np.matrix = None):
-        self.matrixType = matrixType
-        self.values = values
-
-    def from_dict(self, source):
-        if source['_type']:
-            self.matrixType =  source['_type']
-        if source['_values'] and source['_values']['data']:
-            self.values =  np.matrix(source['_values']['data'])
+from app.calculation_service.model.contrast_matrix import ContrastMatrix
 
 
 class OutcomeRepeatedMeasureStDev(object):
